requirements.txt
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
boto3==1.26.48
botocore==1.29.48
cachetools==5.2.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==2.1.1
decorator==4.4.2
exceptiongroup==1.1.0
google-api-core==2.11.0
google-api-python-client==2.72.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
greenlet==2.0.1
httplib2==0.21.0
idna==3.4
imageio==2.24.0
imageio-ffmpeg==0.4.8
iniconfig==2.0.0
jmespath==1.0.1
Markdown==3.4.1
moviepy==2.0.0.dev2
mutagen==1.46.0
numpy==1.24.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
playwright==1.29.1
pluggy==1.0.0
praw==7.6.1
prawcore==2.3.0
proglog==0.1.10
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
pyee==9.0.4
pyparsing==3.0.9
pytest==7.2.0
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-dateutil==2.8.2
python-slugify==7.0.0
pytube==12.1.2
redis==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
text-unidecode==1.3
tinydb==4.7.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.13
websocket-client==1.4.2
